requirements.txt
openai>=1.0.0
python-telegram-bot
redis
python-dotenv
requests
